osetrenie pristupu k storage pre android 6+

+++ app/src/main/java/sk/listok/zssk/zssklistok/helpers/FileHelper.java
@@ -1,8 +1,14 @@
 package sk.listok.zssk.zssklistok.helpers;
 
+import android.Manifest;
 import android.content.Context;
+import android.content.pm.PackageManager;
 import android.content.res.AssetManager;
+import android.os.Build;
 import android.os.Environment;
+import android.support.v4.app.ActivityCompat;
+import android.support.v4.content.ContextCompat;
+import android.support.v7.app.AppCompatActivity;
 import android.util.Log;
 
 import java.io.File;
@@ -24,11 +30,35 @@ public class FileHelper {
     private static Context context;
     private static String dexFileName = "parser.dex";
 
+    /**
+     * Kontrola pre Android 6+, či sú práva na zápis.
+      */
+    private static final int REQUEST_WRITE_STORAGE = 112;
+    private static void checkWriteExternalPermission(AppCompatActivity app){
+        if (Build.VERSION.SDK_INT >= 23) {
+            if (ContextCompat.checkSelfPermission(app, Manifest.permission.WRITE_EXTERNAL_STORAGE)
+                    != PackageManager.PERMISSION_GRANTED) {
+
+                // Should we show an explanation?
+                if (ActivityCompat.shouldShowRequestPermissionRationale(app,
+                        Manifest.permission.WRITE_EXTERNAL_STORAGE)) {
+
+                } else {
+                    ActivityCompat.requestPermissions(app,
+                            new String[]{Manifest.permission.WRITE_EXTERNAL_STORAGE},
+                            REQUEST_WRITE_STORAGE);
+
+                }
+            }
+        }
+    }
+
 
     /**
      * Vráti priečinok kde sú uložené stiahnuté lístky
      */
-    public static File getTicketFolder() {
+    public static File getTicketFolder(AppCompatActivity app) {
+        checkWriteExternalPermission(app);
         String root = Environment.getExternalStorageDirectory().toString();
         File myDir = new File(root + "/vlakoveListky");
         myDir.mkdirs();
@@ -38,7 +68,7 @@ public class FileHelper {
     /**
      * Vráti umiestnenie temp foldera.
      */
-    public static File getTempFolder() {
+    private static File getTempFolder(AppCompatActivity app) {
         String root = context.getCacheDir().toString();
         File myDir = new File(root + "/tmp");
         myDir.mkdirs();
@@ -49,16 +79,17 @@ public class FileHelper {
     /**
      * Vráti .dex súbor(bajtkód) parsera
      */
-    public static File getDexFile() {
+    public static File getDexFile(AppCompatActivity app) {
+        checkWriteExternalPermission(app);
         File dexFile = new File(context.getFilesDir() + "/dexfile/" + dexFileName);
-        dexFile.mkdirs();
         return dexFile;
     }
 
     /**
      * Vráti adresár kde sa .dex nachádza
      */
-    public static File getDexDir() {
+    public static File getDexDir(AppCompatActivity app) {
+        checkWriteExternalPermission(app);
         File dexDir = new File(context.getFilesDir() + "/dexfile");
         dexDir.mkdirs();
         return dexDir;
@@ -68,12 +99,12 @@ public class FileHelper {
     /**
      * Prepíše existujúci .dex súbor, novým bajtkódom.
      */
-    public static void rewriteToDexFile(byte[] bytes) {
+    public static void rewriteToDexFile(byte[] bytes, AppCompatActivity app) {
         try {
-            File f = getDexFile();
+            File f = getDexFile(app);
             f.delete();
 
-            File file = new File(getDexDir(), dexFileName);
+            File file = new File(getDexDir(app), dexFileName);
             FileOutputStream fos = new FileOutputStream(file);
             fos.write(bytes);
             fos.close();
@@ -94,12 +125,13 @@ public class FileHelper {
     /**
      * Skopíruje assets.
      */
-    public static void copyAssets(Context context) {
+    public static void copyAssets(AppCompatActivity context) {
+        checkWriteExternalPermission(context);
         FileHelper.context = context;
 
         //pozriem sa ci v tempe uz nemam
         //nakopirovane subory aby som ich neprevalil
-        File tempDir = FileHelper.getTempFolder();
+        File tempDir = FileHelper.getTempFolder(context);
         ArrayList<String> filesInTemp = new ArrayList<>();
         for (File f : tempDir.listFiles()) {
             if (f.isFile()) {
@@ -128,7 +160,7 @@ public class FileHelper {
             OutputStream out = null;
             try {
                 in = assetManager.open(filename);
-                File outFile = new File(FileHelper.getTempFolder(), filename);
+                File outFile = new File(FileHelper.getTempFolder(context), filename);
                 out = new FileOutputStream(outFile);
                 copyFile(in, out);
             } catch (IOException e) {
@@ -166,7 +198,8 @@ public class FileHelper {
     /**
      * Vymaže lístok z fyzickej pamäte.
      */
-    public static boolean deleteTicket(Ticket ticket) {
+    public static boolean deleteTicket(Ticket ticket, AppCompatActivity app) {
+        checkWriteExternalPermission(app);
         File f = getTicketImage(ticket.getFilename());
         if (f != null) {
             f.delete();
@@ -179,7 +212,11 @@ public class FileHelper {
     /**
      * Vráti priečinok kde sa môže bajtkód nakopírovať, pri loade.
      */
-    public static File getOutputDex(Context mActivity) {
+    public static File getOutputDex(AppCompatActivity mActivity) {
+        checkWriteExternalPermission(mActivity);
         return mActivity.getDir("dex", Context.MODE_PRIVATE);
     }
+
+
+
 }
